requirements.txt
requests
beautifulsoup4
pandas
textract
PyPDF2
spacy
python-dotenv
